% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ABCoptim-package.R
\docType{package}
\name{ABCoptim-package}
\alias{ABCoptim-package}
\alias{ABCoptim}
\alias{abc}
\title{An implementation of the Artificial Bee Colony (ABC) Algorithm}
\description{
This is an implementation of Karaboga (2005) ABC optimization algorithm. It
was developed upon the basic version programmed in \code{C} and distributed
at the algorithm's official website (see the references).
}
\details{
Any evident (precision) error should be blaimed to the package author (not to
the algorithm itself).

Please visit the project home for more information:
https://github.com/gvegayon/ABCoptim.
}
\examples{

  \dontrun{
    demo(ABCoptim) # Some functions...
  }

}
\references{
D. Karaboga, \emph{An Idea based on Honey Bee Swarm for
Numerical Optimization}, tech. report TR06,Erciyes University, Engineering
Faculty, Computer Engineering Department, 2005
http://mf.erciyes.edu.tr/abc/pub/tr06_2005.pdf

Artificial Bee Colony (ABC) Algorithm (website)
http://mf.erciyes.edu.tr/abc/index.htm

Basic version of the algorithm implemented in \code{C} (ABC's official
website) http://mf.erciyes.edu.tr/abc/form.aspx
}
\seealso{
Useful links:
\itemize{
  \item \url{https://gvegayon.github.io/ABCoptim}
  \item \url{https://github.com/gvegayon/ABCoptim}
  \item \url{http://mf.erciyes.edu.tr/abc/}
  \item Report bugs at \url{https://github.com/gvegayon/ABCoptim/issues}
}

}
\author{
\strong{Maintainer}: George Vega Yon \email{g.vegayon@gmail.com} (\href{https://orcid.org/0000-0002-3171-0844}{ORCID})

Authors:
\itemize{
  \item George Vega Yon \email{g.vegayon@gmail.com} (\href{https://orcid.org/0000-0002-3171-0844}{ORCID})
}

Other contributors:
\itemize{
  \item Enyelbert Muñoz \email{enyeldoc2011@gmail.com} [contributor]
}

}
\keyword{package}
