\name{ABCoptim-package}
\alias{ABCoptim-package}
\alias{ABCoptim}
\alias{abc}
\docType{package}
\title{
  An implementation of the Artificial Bee Colony (ABC) Algorithm
}
\description{
  Please visit the project home for more information:
    \url{https://github.com/gvegayon/ABCoptim}.
} 
\details{
  \tabular{ll}{
    Package: \tab ABCoptim\cr
    Type: \tab Package\cr
    Version: \tab 0.13.10\cr
    Date: \tab 2013-10-20\cr
    License: \tab GPL version 2 or later\cr
  }
}
\author{George Vega Yon \email{g.vegayon@gmail.com} [aut],

Enyelbert Mu\~noz \email{enyeldoc2011@gmail.com} [cnt)}

\examples{
  \dontrun{
    demo(ABCoptim) # Some functions...
  }
}
\keyword{ package }
