% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ABCoptim-package.R
\docType{package}
\name{ABCoptim-package}
\alias{ABCoptim-package}
\alias{ABCoptim}
\alias{abc}
\title{An implementation of the Artificial Bee Colony (ABC) Algorithm}
\description{
This is an implementation of Karaboga (2005) ABC optimization algorithm. It
was developed upon the basic version programmed in \code{C} and distributed
at the algorithm's official website (see the references).
}
\details{
Any evident (precision) error should be blaimed to the package author (not to
the algorithm itself).

Please visit the project home for more information:
https://github.com/gvegayon/ABCoptim.
}
\examples{

  \dontrun{
    demo(ABCoptim) # Some functions...
  }

}
\references{
D. Karaboga, \emph{An Idea based on Honey Bee Swarm for
Numerical Optimization}, tech. report TR06,Erciyes University, Engineering
Faculty, Computer Engineering Department, 2005
http://mf.erciyes.edu.tr/abc/pub/tr06_2005.pdf

Artificial Bee Colony (ABC) Algorithm (website)
http://mf.erciyes.edu.tr/abc/index.htm

Basic version of the algorithm implemented in \code{C} (ABC's official
website) http://mf.erciyes.edu.tr/abc/form.aspx
}
\keyword{package}
\section{How to cite}{
  If you use \pkg{ABCoptim} in published work, please cite it. Run
  \code{citation("ABCoptim")} in R for the full entry.
}
