\name{ABCoptim-package}
\alias{ABCoptim-package}
\alias{ABCoptim}
\alias{abc}
\docType{package}
\title{
  An implementation of the Artificial Bee Colony (ABC) Algorithm
}
\description{
  Please visit the project home for more information:
    \url{https://bitbucket.org/gvegayon/rgexf}.
} 
\details{
  \tabular{ll}{
    Package: \tab abc_optimization\cr
    Type: \tab Package\cr
    Version: \tab 0.13.10\cr
    Date: \tab 2013-10-20\cr
    License: \tab GPL version 2 or later\cr
  }
}
\author{George Vega Yon \email{george.vega@nodoschile.org} [aut],

Jorge Fabrega Lacoa \email{jorge.fabrega@nodoschile.org} [cnt),

Joshua B. Kunst \email{jbkunst@gmail.com} [cnt]}
\note{
  See the GEXF primer for details on the GEXF graph format:
  \url{http://gexf.net/1.2draft/gexf-12draft-primer.pdf}
}
\references{
  \itemize{
    \item rgexf project site: \url{https://bitbucket.org/gvegayon/rgexf}
    \item Gephi project site: \url{https://gephi.org/}
    \item GEXF project site: \url{http://gexf.net/}
    \item Nodos Chile project site (real life examples): \url{http://nodoschile.org/}
    \item Sigmasj project site : \url{http://sigmajs.org/}
  }
}
\examples{
  \dontrun{
    demo(gexf) # Example of gexf command using fictional data.
    demo(gexfattributes) # Working with attributes.
    demo(gexfbasic) # Basic net.
    demo(gexfdynamic) # Dynamic net.
    demo(edge.list) # Working with edges lists.
    demo(gexffull) # All the package.
    demo(gexftwitter) # Example with real data of chilean twitter accounts.
    demo(gexfdynamicandatt) # Dynamic net with static attributes.
    demo(gexfbuildfromscratch) # Example building a net from scratch.
    demo(gexfigraph) # Two-way gexf-igraph conversion
  }
}
\keyword{ package }
